refactor: rename directory name from preprocess to process
